Pin to max Python 3.12, move around some ruff config
Note that the Docker image we use has Python 3.11, so functions will
always run with that version regardless of whether you have 3.12
locally. (Unless you choose to use a different base image.)

The ruff config tweaks are because they moved some config around with a
newer version that was updated by renovate a while back.

Signed-off-by: Nic Cope <nicc@rk0n.org>